Tornado Blog/Blog_Project/requirements.txt
backports-abc==0.5
backports.ssl-match-hostname==3.5.0.1
beautifulsoup4==4.6.0
bs4==0.0.1
certifi==2015.4.28
chardet==3.0.4
idna==2.5
itchat==1.3.9
olefile==0.44
Pillow==4.1.1
pycket==0.3.0
PyMySQL==0.7.11
pypng==0.0.18
PyQRCode==1.2.1
qrcode==5.3
redis==2.10.5
requests==2.18.1
singledispatch==3.4.0.3
six==1.10.0
SQLAlchemy==1.1.11
tornado==4.5.1
urllib3==1.21.1


